# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sajamaax",
        "description": "api Sajaamax",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://51.77.94.252:8080",
            "description": "Sajamaax address"
        }
    ],
    "paths": {
...
        "/api/user/update-password": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update pass",
                "description": "update pass user",
                "operationId": "updateFeedbackItem",
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "422": {
                        "description": "Data Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}